Fold commandContext and clusterContext back into starrocks.go
Keep job_context.go and backend_client_pool.go split out, but
commandContext/clusterContext (plugin.Handler entry points +
cluster config) stay in the main file.
